234A is complete.
This program is locked with a user and a pass. If a file does not exist
to tell these details to the program, a file will be created and details
will be asked for. If a file does exist, then the user has three
attempts to open the file with the proper inputs.

I wanted to try out classes again. I had trouble with referencing class
instances, since I used multiple functions. So, I defined the
troublesome function as a local function, and referenced the class parts
within the main function. It works just fine, although I probably didn't
need to put the function in as a local.

I forgot to move the comments back after I put them outside main, so
i'll fix that for tomorrow.

In the future, i'd like to make a program that would open the file for
the user if they forgot the user and pass. I'm thinking that looping
over a series of letters via ordinal numbers should do the trick.