94A is complete - Monty Hall Problem
This program simulates the Monty Hall Problem where you have three
doors, where a prize is behind one of them. You select a door, and
someone opens one door to reveal that it is empty, before asking if you
want to switch. This program reveals the chance of winning by staying
and the chance of winning by switching.