7A is complete
This program consumes two integer inputs and produces the dottie
number(a number produced by recursing cos(x) until the value stays the
same), the results of recursing x - tan(x) (with first x = 2) and the
result recursing 1 - 1/x with a given x until a value stays the same.

This was an interesting problem to work on. The tan problem goes to pi,
and 1 - 1/x goes to the golden ratio. It's interesting to learn about
these kinds of mathematics.