14A is complete
This program accepts a string input (all caps) and returns that string
in its "balanced form". To balance a string, a letter is selected as a
balance point. The balance point has a position in the string. The
further any other letter is from that balance point, the more of a
multiplier it gets (right next to balance = 1x, 3rd letter from balance
= 3x). Every letter of the alphabet has a value (A = 1,..., Z = 26). The
word is balanced is the total letter values to the left of the balance
point equal total letter values of the right of the balance point
(including multipliers for each side). Some words are not balanceable,
and the program will tell you of those words.

I want to update this program with a dictionary that is defined by a
loop. I tried to do that earlier, but it didn't work, so I had to define
the dictionary manually. I should come back to this when i'm a little
better at python and make it better.