18A is done - palindromic number generator
This program accepts an integer and returns that integer in palindromic
form (read the same from front to back). If the number is not already
palindromic, then the digits are reversed, and this new number is added
to the original number. If the sum is not palindromic, the process is
repeated until it is palindromic. The number of steps are also
outputted.

I liked this program, since I didn't know that this process could lead
to palindromic numbers. Maybe this program could be used in a later
project, like an encryption algorithm.