peopleInSpace is done - tells who is in space at current moment
This was just a mini project to work with the requests module and
working with jsons.
This program will tell you how many people are in space right now, along
with who those people are. It's really short, but this was just to test
out how to use requests and how to use jsons in python 3