seriesWin is done - tree probability finder
This program takes two inputs: the number of games in a series, and the
chance that team A has for winning one game. This program will find the
chance that team A wins the series, assuming they are facing team B who
has 1 - prob(A) of winning a game. They win the series if they win the
majority of the games that are played in the series (ex. if 7-game
series, 4 must be won).

This program analyzes all outcomes that occur, and outputs the sum of
the probabilities of all favourable outcomes.