154B is in progress
This program takes in a yyyy,mm,dd date format and outputs what day of
the week that day is. This uses the doomsday algorithm.

The problem i've been facing involved string indicies (because I used a
comma instead of a colon and never noticed it), but now it appears to be
a logic error. It claims that Apr, 05/ 1995 is a Tuesday, when it
actually was a Wednesday.